Fix clusterize not working in container
Clusterize needs a special set up inside the singularity container
(e.g., mounting /opt/slurm-skretch{,-testing}/deployed/etc to
/etc/slurm-config).

We track that the setup is present via `${CLUSTERIZE_SET_UP}`, if it is
not, we perform the regular singularity call even if we already are in a
container.

This fixes problems with jenkins jobs that are run in a container but do
not have the clusterize set up.

Change-Id: I467bde5e56bdba3f7037a3fc883a624ab9cfb270