clusterize: Add alternative mount point to add_if_exists
Allow for a third argument to `add_if_exists` that gets added if the
original mount point does not exists.

This is needed for nested usage of clusterize in testing deployments
where, in the outter layer, we need to map /run/munge-testing to
/run/munge inside the container, but once we are inside the container
and call clusterize again (e.g., for running quiggeldy or a user
script that is interacting with quiggeldy), /run/munge-testing does not
exist anymore and /run/munge should be binded into the next level as if
we are in the production environment.

Change-Id: I7ee50a9e027841741085fee3d8ef7ec6862e8971