clusterize: Support nesting of clusterize calls
Add support for clusterize to be called from within a clusterize'd
environment. This is useful when starting scripts and
services that rely on a functioning slurm environment. For instance,
quiggeldy needs munge-based authentication to verify users and
facilitate fair-sharing between them.

For this the following additions had to be implemented:
* Unset CLUSTERIZE_SET_UP after we have set up the environment in the
  container.
* Make sure that we do not add two bind-requests for /opt/slurm if we
  are already within the container (will lead to error).

Change-Id: I2cbb5d520daac04ee92bbac7a088fb666d852177