clusterize: Fix PATH handling
If "default" folders such as /usr/local/bin are present in the $PATH
variable, they get put _ahead_ of app-specific folders in the
container. This leads to shadowing of app-specific binaries (such as
python) by their global counterparts (i.e., the version of python that
was used during installation).

The solution is the purge these common paths from $PATH when escaping
the environment as singularity will append them.

Change-Id: I5464d2cd79eea59958c8c4610a20bf755082068a