clusterize: Fix error with trailing PATH-entries
Because the the sub-function from gawk does not support backreferences,
the leading colon is not re-inserted when the default paths to purge are
trailed by any other paths, rendering the remaining path useless.

We fix this by creating a single gensub-call (with some bash magic that
properly escapes and joins the paths to be used in the regex) that
leaves colons untouched.

A second substitution then replaces multiple consecutive colons with a
single one.

Change-Id: I4f145aaeb62ac4f3c351a5cafd671a94003fc53e